3 - move flip tile utils
