+++ src/utils/flipSplashArea.js
